# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz App",
        "description": "Documentación de mi API REST",
        "contact": {
            "email": "kvfa13@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/questions/{id}/player-answer": {
            "post": {
                "tags": [
                    "Jogo"
                ],
                "summary": "Registrar resposta do jogador para uma pergunta",
                "description": "Permite que um jogador registre sua resposta para uma pergunta específica. O usuário não pode responder a pergunta mais de uma vez.",
                "operationId": "08e8df2f1b083a20a5e543e390b30867",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pergunta para a qual o jogador está respondendo.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "given_answer": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resposta do jogador registrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "question_id": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "given_answer": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-29T09:50:05.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-29T09:50:05.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 13
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao registrar resposta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have already answered this question."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pergunta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No results found for Question with ID 25"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/my-answer-question": {
            "get": {
                "tags": [
                    "Jogo"
                ],
                "summary": "Obter as respostas do usuário autenticado",
                "description": "Este endpoint permite que usuários autenticados obtenham uma lista de suas respostas às perguntas, incluindo se foram corretas ou incorretas.",
                "operationId": "getUserAnswers",
                "responses": {
                    "200": {
                        "description": "Lista de respostas do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Identificador único da resposta",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "description": "Identificador do usuário que respondeu à pergunta",
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "is_correct": {
                                                        "description": "Indica se a resposta é correta (1) ou incorreta (0)",
                                                        "type": "boolean",
                                                        "example": 0
                                                    },
                                                    "question_name": {
                                                        "description": "Nome ou texto da pergunta",
                                                        "type": "string",
                                                        "example": "pergunta falsa no projeto"
                                                    },
                                                    "question_image": {
                                                        "description": "URL da imagem associada à pergunta, se existir",
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado - O usuário não está autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}/answers": {
            "get": {
                "tags": [
                    "Jogo"
                ],
                "summary": "Obter respostas de um usuário específico",
                "description": "Este endpoint permite obter uma lista de respostas de um usuário específico, identificando-o pelo seu ID na URL.",
                "operationId": "getUserAnswersById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário cujas respostas se deseja obter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de respostas do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Identificador único da resposta",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "description": "ID do usuário que respondeu à pergunta",
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "is_correct": {
                                                        "description": "Indica se a resposta é correta (1) ou incorreta (0)",
                                                        "type": "boolean",
                                                        "example": 0
                                                    },
                                                    "question_name": {
                                                        "description": "Nome ou texto da pergunta",
                                                        "type": "string",
                                                        "example": "Pergunta de exemplo"
                                                    },
                                                    "question_image": {
                                                        "description": "URL da imagem associada à pergunta, se existir",
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado - O usuário não está autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuário não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}